requirements.txt
Django==1.8.1
docutils==0.12
gunicorn==19.3.0
Pillow==2.8.1
psycopg2==2.6
reportlab==3.1.44
uWSGI==2.0.10
validate-email==1.3
virtualenv==13.1.0
